Se agregaron las comprobaciones de las tablas para que no de error al hacer migraciones

+++ barberia/database/migrations/2024_10_28_185759_create_cita_servicio_table.php
@@ -8,19 +8,21 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('cita_servicio', function (Blueprint $table) {
-            $table->unsignedBigInteger('id_cita');
-            $table->foreign('id_cita')->references('id_cita')->on('cita')->onDelete('cascade')->onUpdate('cascade');
-            
-            $table->unsignedBigInteger('id_servicio');
-            $table->foreign('id_servicio')->references('id_servicio')->on('servicio')->onDelete('restrict')->onUpdate('cascade');
-            
-            $table->integer('cantidad')->default(1);
-            $table->decimal('costo', 10, 2);
-            $table->integer('duracion');
-            $table->primary(['id_cita', 'id_servicio']); // Clave primaria compuesta
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('cita_servicio')) {
+            Schema::create('cita_servicio', function (Blueprint $table) {
+                $table->unsignedBigInteger('id_cita');
+                $table->foreign('id_cita')->references('id_cita')->on('cita')->onDelete('cascade')->onUpdate('cascade');
+                
+                $table->unsignedBigInteger('id_servicio');
+                $table->foreign('id_servicio')->references('id_servicio')->on('servicio')->onDelete('restrict')->onUpdate('cascade');
+                
+                $table->integer('cantidad')->default(1);
+                $table->decimal('costo', 10, 2);
+                $table->integer('duracion');
+                $table->primary(['id_cita', 'id_servicio']); // Clave primaria compuesta
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
